Id values fixed Long to Integers. Closes #12

+++ pet_clinic_data/src/main/java/tim/spring/app/sft_pet_clinic/services/OwnerService.java
@@ -4,7 +4,7 @@ import tim.spring.app.sft_pet_clinic.model.Owner;
 
 import java.util.Set;
 
-public interface OwnerService extends CrudService<Owner,Long>{
+public interface OwnerService extends CrudService<Owner,Integer>{
 
     Set<Owner> findByLastName(String lastName);
 }
